pyproject.toml
[project]
name = "etl-monitor"
version = "0.1.0"
description = "Multi-Agent ETL Monitor & Self-Healer using AutoGen + Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autogen-agentchat>=0.4",
    "autogen-ext[openai]>=0.4",
    "pandas>=2.2",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/etl_monitor"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
